Separate external secondary instance parse tests
No functional changes, only rearranging
